app/requirements.txt
Flask
pandas
mysql-connector
pymysql
sqlalchemy